Moved badger route state selectors in spearate file as per review comments
